requirements.txt
PyPDF2==3.0.1
google-generativeai==0.8.5
tiktoken==0.5.1
scikit-learn==1.3.0
numpy==1.24.3
streamlit==1.29.0
python-dotenv==1.0.0